requirements.txt
lxml==4.9.3
requests==2.31.0
ruamel.yaml==0.17.35